feat:update TTS voice diy and update LLMFactory

+++ src/main/java/top/rslly/iot/services/thingsModel/ProductServiceImpl.java
@@ -25,7 +25,6 @@ import org.springframework.transaction.annotation.Transactional;
 import top.rslly.iot.dao.*;
 import top.rslly.iot.models.*;
 import top.rslly.iot.param.request.Product;
-import top.rslly.iot.param.request.ProductToolsBan;
 import top.rslly.iot.utility.JwtTokenUtil;
 import top.rslly.iot.utility.result.JsonResult;
 import top.rslly.iot.utility.result.ResultCode;
@@ -67,6 +66,8 @@ public class ProductServiceImpl implements ProductService {
   private ProductToolsBanRepository productToolsBanRepository;
   @Resource
   private AgentLongMemoryRepository agentLongMemoryRepository;
+  @Resource
+  private ProductVoiceDiyRepository productVoiceDiyRepository;
 
 
   @Override
@@ -184,6 +185,8 @@ public class ProductServiceImpl implements ProductService {
         productToolsBanRepository.findAllByProductId(id);
     List<AgentLongMemoryEntity> agentLongMemoryEntityList =
         agentLongMemoryRepository.findAllByProductId(id);
+    List<ProductVoiceDiyEntity> productVoiceDiyEntityList =
+        productVoiceDiyRepository.findAllByProductId(id);
     boolean p1 = productModelEntityList.isEmpty();
     boolean p2 = wxProductBindEntityList.isEmpty();
     boolean p3 = otaEntityList.isEmpty();
@@ -193,7 +196,8 @@ public class ProductServiceImpl implements ProductService {
     boolean p7 = productRouterSetEntityList.isEmpty();
     boolean p8 = productToolsBanEntityList.isEmpty();
     boolean p9 = agentLongMemoryEntityList.isEmpty();
-    if (p1 && p2 && p3 && p4 && p5 && p6 && p7 && p8 && p9) {
+    boolean p10 = productVoiceDiyEntityList.isEmpty();
+    if (p1 && p2 && p3 && p4 && p5 && p6 && p7 && p8 && p9 && p10) {
       List<ProductEntity> result = productRepository.deleteById(id);
       if (result.isEmpty())
         return ResultTool.fail(ResultCode.PARAM_NOT_VALID);
